requirements.txt
pytesseract
pillow
numpy
opencv-python